requirements.txt
streamlit
streamlit-lottie
langchain-google-genai
langchain-core
langchain-community
pydantic
python-dotenv
google-generativeai
geopy
requests
Pillow
plotly
pandas
bcrypt
pygame
streamlit-modal
streamlit_js_eval
scikit-learn
numpy
opencv-python
mediapipe==0.10.14
tensorflow

